+++ packages/app/components/header/header-right.tsx
@@ -2,86 +2,21 @@ import { useWindowDimensions } from "react-native";
 
 import { Button } from "@showtime-xyz/universal.button";
 import { useIsDarkMode } from "@showtime-xyz/universal.hooks";
-import { ArrowLeft, Plus, Search } from "@showtime-xyz/universal.icon";
+import { Plus } from "@showtime-xyz/universal.icon";
 import { PressableHover } from "@showtime-xyz/universal.pressable-hover";
-import { PressableScale } from "@showtime-xyz/universal.pressable-scale";
 import { useRouter } from "@showtime-xyz/universal.router";
 import { View } from "@showtime-xyz/universal.view";
 
-// import { NetworkButton } from "app/components/connect-button";
 import { HeaderDropdown } from "app/components/header-dropdown";
 import { useRedirectToCreateDrop } from "app/hooks/use-redirect-to-create-drop";
 import { useUser } from "app/hooks/use-user";
-import {
-  ShowtimeTabBarIcon,
-  TrendingTabBarIcon,
-} from "app/navigation/tab-bar-icons";
+import { TrendingTabBarIcon } from "app/navigation/tab-bar-icons";
 import { useNavigateToLogin } from "app/navigation/use-navigate-to";
 
 import { breakpoints } from "design-system/theme";
 
-import { NotificationsInHeader, SearchInHeader } from "./header";
+import { NotificationsInHeader } from "./header";
 
-export * from "./header";
-type HeaderLeftProps = {
-  canGoBack: boolean;
-  withBackground?: boolean;
-  color?: string;
-};
-export const HeaderLeft = ({
-  canGoBack,
-  withBackground = false,
-  color,
-}: HeaderLeftProps) => {
-  const isDark = useIsDarkMode();
-  const router = useRouter();
-  const Icon = canGoBack ? ArrowLeft : Search;
-
-  return (
-    <PressableScale
-      hitSlop={{ top: 10, bottom: 10, left: 10, right: 10 }}
-      style={[
-        {
-          height: 32,
-          width: 32,
-          justifyContent: "center",
-          alignItems: "center",
-          borderRadius: 999,
-        },
-        withBackground && { backgroundColor: "rgba(0,0,0,.3)" },
-      ]}
-      onPress={() => {
-        if (canGoBack) {
-          router.pop();
-        } else {
-          router.push("/search");
-        }
-      }}
-    >
-      <Icon
-        color={
-          color ? color : withBackground ? "#FFF" : isDark ? "#FFF" : "#000"
-        }
-        width={24}
-        height={24}
-      />
-    </PressableScale>
-  );
-};
-export const HeaderCenter = ({
-  isDark,
-  isMdWidth,
-}: {
-  isDark?: boolean;
-  isMdWidth?: boolean;
-}) => {
-  return (
-    <View tw="flex flex-row">
-      <ShowtimeTabBarIcon color={isDark ? "black" : "white"} tw="mr-4" />
-      {isMdWidth ? <SearchInHeader /> : null}
-    </View>
-  );
-};
 export const HeaderRight = () => {
   const router = useRouter();
   const { isLoading, isAuthenticated } = useUser();
