+++ packages/app/components/header/header.tsx
@@ -11,7 +11,7 @@ import { useRouter } from "@showtime-xyz/universal.router";
 import { useSafeAreaInsets } from "@showtime-xyz/universal.safe-area";
 import { View } from "@showtime-xyz/universal.view";
 
-import { HeaderLeft } from ".";
+import { HeaderLeft } from "./header-left";
 
 export const DEFAULT_HADER_HEIGHT = 44;
 const renderComponent = (Component: any) => {
