+++ packages/app/components/header/index.ts
@@ -1,0 +1,4 @@
+export * from "./header";
+export * from "./header-center";
+export * from "./header-left";
+export * from "./header-right";
