+++ packages/app/components/settings/index.ts
@@ -1,28 +1,1 @@
 export { Settings } from "./setting";
-export const SETTINGS_ROUTES = [
-  {
-    title: "Wallets",
-    key: "Wallets",
-    index: 0,
-  },
-  {
-    title: "Email",
-    key: "Email",
-    index: 1,
-  },
-  {
-    title: "Phone Number",
-    key: "Phone",
-    index: 2,
-  },
-  {
-    title: "Account",
-    key: "Account",
-    index: 3,
-  },
-  {
-    title: "Push Notifications",
-    key: "Push Notifications",
-    index: 4,
-  },
-];
