+++ packages/app/lib/constants.ts
@@ -7,6 +7,7 @@ export const colors = {
   pink: "#e45cff",
   teal: "#1dd4e0",
 };
+export const BOTTOM_TABBAR_BASE_HEIGHT = 64;
 
 export const CONTRACTS = {
   ZORA: "0xabefbc9fd2f806065b4f3c237d4b59d9a97bcac7",
